[Refactor] new base class: `Layoutable`
